updated display
Made display the controller class for the JFrame and allowed panels to be added and subtracted to/from it for different types of screens. Added PicPanel a type of JPanel that allows image backgrounds and resizing of those images. Travel Screen is the same as what was created for MVP #1.